Create class Finder, to list files for download
